Nectar Betta - not final
There is an uresolved "bug" - using this versin with Modularity gives
slightly different results in comparison with the first version. It is
to be fixed.
Comparing them, the first partition and metaData are the same. But after
the first iteration there are different commms.
Generate output dir if dont exist